Рабочая версия

+++ todo_list_winforms/Presenters/TodoNotePresenter.cs
@@ -1,8 +1,9 @@
-﻿using System.Runtime.CompilerServices;
+﻿using System.Timers;
 using todo_list_winforms.DAL.Models;
 using todo_list_winforms.DAL.Repository;
 using todo_list_winforms.Migrations;
 using todo_list_winforms.Views;
+using Timer = System.Timers.Timer;
 
 namespace todo_list_winforms.Presenters;
 
@@ -11,26 +12,42 @@ public class TodoNotePresenter
     private ITodoNoteView _view;
     private ITodoNoteRepository _repository;
     private BindingSource _bindingSource;
-    private IEnumerable<TodoNote> _noteList;
+    private IEnumerable<TodoNote> _todoList;
 
     public TodoNotePresenter(ITodoNoteView view, ITodoNoteRepository repository)
     {
         _view = view;
         _repository = repository;
         _bindingSource = new();
-        _view.EditEvent += LoadSelectedTodoToEdit;
+        _view.EditEvent += SelectTodo;
         _view.DeleteEvent += DeleteSelectedTodo;
-        _view.CreateEvent += SaveTodo;
+        _view.CreateEvent += CreateTodo;
+        _view.UpdateEvent += UpdateTodo;
         _view.CancelEvent += CancelAction;
+        _view.OnlyCompletedEvent += LoadNotCompletedTodoList;
+        _view.TodoListClickEvent += SelectTodo;
         _view.SetTodoListBindingSource(_bindingSource);
         LoadAllTodoList();
         _view.Show();
     }
 
-    private void LoadAllTodoList()
+    private void LoadNotCompletedTodoList(object? sender, EventArgs e) => LoadAllTodoList(_view.ShowOnlyCompleted);
+    private void LoadAllTodoList(bool notCompletedOnly = false)
     {
-        _noteList = _repository.GetAll();
-        _bindingSource.DataSource = _noteList;
+        _todoList = _repository.GetAll(notCompletedOnly);
+
+        foreach (var todo in _todoList)
+        {
+            if (todo.EndDate == null) continue;
+            TimeSpan timeUntilEvent = todo.EndDate.Value.Subtract(DateTime.Now).Subtract(TimeSpan.FromMinutes(5));
+            if (timeUntilEvent.TotalMilliseconds <= 0 || timeUntilEvent.TotalMilliseconds >= double.MaxValue || (int)timeUntilEvent.TotalMilliseconds<0) continue;
+            var timer = new Timer(timeUntilEvent.TotalMilliseconds);
+            timer.Elapsed += NotifySoonExpires;
+            timer.AutoReset = false;
+            timer.Start();
+        }
+
+        _bindingSource.DataSource = _todoList;
     }
 
     private void CancelAction(object? sender, EventArgs e)
@@ -38,28 +55,34 @@ public class TodoNotePresenter
         CleanViewFields();
     }
 
-    private void SaveTodo(object? sender, EventArgs e)
+    private void UpdateTodo(object? sender, EventArgs e)
     {
-        var model = new TodoNote();
-        model.Id = _view.Id;
-        model.Title = _view.Title;
-        model.Description = _view.Description;
-        model.IsDone = _view.IsDone;
-        model.StartDate = _view.StartDate;
-        model.EndDate = _view.EndDate;
+        var model = MapViewToModel();
+        if (model == null || model.Id == 0) { return; }
+        try
+        {
+            _repository.Update(model);
+            _view.Message = "Задача обновлена";
+            _view.IsSuccessful = true;
+            LoadAllTodoList();
+        }
+        catch (Exception ex)
+        {
+            _view.IsSuccessful = false;
+            _view.Message = ex.Message;
+            throw;
+        }
+    }
+
+    private void CreateTodo(object? sender, EventArgs e)
+    {
+        var model = MapViewToModel();
+        model.Id = 0;
 
         try
         {
-            if (_view.IsEdit)
-            {
-                _repository.Update(model);
-                _view.Message = "Ура обновлено";
-            }
-            else
-            {
-                _repository.Create(model);
-                _view.Message = "Ура добавлено";
-            }
+            _repository.Create(model);
+            _view.Message = "Задача создана";
             _view.IsSuccessful = true;
             LoadAllTodoList();
             CleanViewFields();
@@ -72,12 +95,16 @@ public class TodoNotePresenter
         }
     }
 
+    private void NotifySoonExpires(object sender, ElapsedEventArgs e)
+    {
+        _view.ShowMessage("Через 5 минут истечет крайний срок задачи!");
+    }
+
     private void CleanViewFields()
     {
         _view.Id = 0;
         _view.Title = "";
         _view.Description = "";
-        _view.IsEdit = false;
         _view.StartDate = null;
         _view.EndDate = null;
     }
@@ -87,27 +114,42 @@ public class TodoNotePresenter
         try
         {
             var todo = (TodoNote)_bindingSource.Current;
+            if(todo==null || todo.Id == 0) { return; }
             _repository.Delete(todo.Id);
             _view.IsSuccessful = true;
-            _view.Message = "Удалилось";
+            _view.Message = "Задача удалена";
+            CleanViewFields();
             LoadAllTodoList();
         }
         catch (Exception)
         {
             _view.IsSuccessful = false;
-            _view.Message = "Error БЛЯТЬ!";
+            _view.Message = "Ошибка!";
             throw;
         }
     }
 
-    private void LoadSelectedTodoToEdit(object? sender, EventArgs e)
+    private void SelectTodo(object? sender, EventArgs e)
     {
         var todo = (TodoNote)_bindingSource.Current;
         _view.Id = todo.Id;
         _view.Title = todo.Title;
         _view.Description = todo.Description;
-        _view.IsEdit = true;
-        _view.StartDate = DateTime.Now;
+        _view.StartDate = todo.StartDate;
         _view.EndDate = todo.EndDate;
+        _view.IsDone = todo.IsDone;
+    }
+
+    private TodoNote MapViewToModel()
+    {
+        return new TodoNote
+        {
+            Id = _view.Id,
+            Title = _view.Title,
+            Description = _view.Description,
+            IsDone = _view.IsDone,
+            StartDate = _view.StartDate,
+            EndDate = _view.EndDate
+        };
     }
 }
